extract function to get multiple secrets from sm
