+++ pkg/resolvers/secretsmanager.go
@@ -3,11 +3,8 @@ package resolvers
 import (
 	"context"
 	"fmt"
-	"os"
 	"regexp"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/aws/retry"
@@ -16,54 +13,6 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/secretsmanager/types"
 	"github.com/roverdotcom/snagsby/pkg/config"
 )
-
-var smConcurrency int
-
-func init() {
-	// Pull concurrency settings
-	getConcurrency, hasSetting := os.LookupEnv("SNAGSBY_SM_CONCURRENCY")
-	if hasSetting {
-		i, err := strconv.Atoi(getConcurrency)
-		if err == nil && i >= 0 {
-			smConcurrency = i
-		}
-	}
-}
-
-// Concurrency work
-type smMessage struct {
-	Source      *config.Source
-	Name        *string
-	Result      string
-	Error       error
-	IsRecursive bool
-}
-
-func smWorker(jobs <-chan *smMessage, results chan<- *smMessage, svc *secretsmanager.Client) {
-	for job := range jobs {
-		sourceURL := job.Source.URL
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-		input := &secretsmanager.GetSecretValueInput{
-			SecretId: job.Name,
-		}
-		versionStage := sourceURL.Query().Get("version-stage")
-		if versionStage != "" {
-			input.VersionStage = aws.String(versionStage)
-		}
-		versionID := sourceURL.Query().Get("version-id")
-		if versionID != "" {
-			input.VersionId = aws.String(versionID)
-		}
-		getSecret, err := svc.GetSecretValue(ctx, input)
-		if err != nil {
-			job.Error = err
-		} else {
-			job.Result = *getSecret.SecretString
-		}
-		results <- job
-	}
-}
 
 // SecretsManagerResolver handles secrets manager resolution
 type SecretsManagerResolver struct{}
@@ -119,39 +68,7 @@ func (s *SecretsManagerResolver) resolveRecursive(source *config.Source) *Result
 
 	}
 
-	jobs := make(chan *smMessage, len(secretKeys))
-	results := make(chan *smMessage, len(secretKeys))
-
-	// Determine concurrency level defaulting to number of secrets to pull
-	var numWorkers int
-	if smConcurrency > 0 {
-		numWorkers = smConcurrency
-	} else {
-		numWorkers = len(secretKeys)
-	}
-
-	// Boot up workers
-	for w := 1; w <= numWorkers; w++ {
-		go smWorker(jobs, results, svc)
-	}
-
-	// Publish to workers
-	for _, name := range secretKeys {
-		jobs <- &smMessage{Source: source, Name: name}
-	}
-	close(jobs)
-
-	// Loop through results
-	for a := 1; a <= len(secretKeys); a++ {
-		res := <-results
-		if res.Error != nil {
-			result.AppendError(res.Error)
-		} else {
-			result.AppendItem(s.keyNameFromPrefix(prefix, *res.Name), res.Result)
-		}
-	}
-
-	return result
+	return getSecrets(source, svc, secretKeys)
 }
 
 func (s *SecretsManagerResolver) resolveSingle(source *config.Source) *Result {
